renamed: 파일 이동 (components/modal/hook/useScrollLock.ts -> hooks/useScrollLock.ts) - Issue #47
